Move the validation logic to the worship service class
(Not relevant for central worship services).
