+++ app/models/worship-administrative-unit.js
@@ -9,7 +9,6 @@ import {
   WorshipServiceCodeList,
 } from '../constants/Classification';
 import { CLASSIFICATION } from './administrative-unit-classification-code';
-import Joi from 'joi';
 
 export default class WorshipAdministrativeUnitModel extends AdministrativeUnitModel {
   @belongsTo('recognized-worship-type', {
@@ -42,25 +41,7 @@ export default class WorshipAdministrativeUnitModel extends AdministrativeUnitMo
         REQUIRED_MESSAGE,
       ),
       ministerPositions: validateHasManyOptional(),
-      involvements: Joi.when(Joi.ref('$involvementsPercentage'), {
-        is: Joi.exist().valid(true),
-        then: Joi.array().external(async (_value, helpers) => {
-          const involvements = await this.involvements;
-
-          const sumPercentages = involvements.reduce(
-            (percentageAcc, involvement) =>
-              percentageAcc + Number(involvement.percentage),
-            0,
-          );
-
-          if (Number.isNaN(sumPercentages) || sumPercentages !== 100) {
-            return helpers.message(
-              'Het totaal van alle percentages moet gelijk zijn aan 100',
-            );
-          }
-        }),
-        otherwise: validateHasManyOptional(),
-      }),
+      involvements: validateHasManyOptional(),
     });
   }
